+++ doc/conf.py
@@ -19,7 +19,6 @@ import os
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('../'))
 
-from sickle import __version__ as version
 
 # -- General configuration -----------------------------------------------
 
@@ -52,7 +51,7 @@ copyright = u'2013, Mathias Loesch'
 # built documents.
 #
 # The short X.Y version.
-version = version
+version = '0.3'
 # The full version, including alpha/beta/rc tags.
 release = version
 
@@ -124,7 +123,7 @@ html_logo = 'sickle_logo.png'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+# html_static_path = ['_static']
 
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
 # using the given strftime format.